server/requirements.txt
torch==1.6.0
Flask==1.1.2
networkx==2.4
pandas==1.1.1
numpy==1.19.1
requests==2.24.0
transformers==3.3.1
flask_cors==3.0.9
python-dotenv==0.14.0
rake_nltk==1.0.4
